+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
+## Dev
+
+- **args are changed to RenderingContext for rendering steps
+- **args are changed to BindingContext for binding
+
 ## 2.2.0
 
 - Expression resolved using code as resolve parameter
 - Migrated to injectool 1.1.1
-- **args are changed to RenderingContext for rendering steps
 
 ## 2.1.1